requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-jazzmin==2.4.9
django-suit==0.2.26
djangorestframework==3.11.1
exchange==0.3
forex-python==1.5
idna==2.10
openexchangerates==0.1.1
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
xlwt==1.3.0
